Fixed Build failure and compilation error for examples/* / *.ino

+++ examples/i2c_command/i2c_command.ino
@@ -23,9 +23,9 @@ void setup()
 
 void loop()
 {
-  int32_t temperature;
-  int32_t pressure;
-  int16_t oversampling = 7;
+  float temperature;
+  float pressure;
+  uint8_t oversampling = 7;
   int16_t ret;
   Serial.println();
